Generic Code cleanup
- Use better boolean logic
- Reformat and cleanup comments
- Start Javadoc documentation
- Proper packaging (source packages)

This is intended as a learning exercise, and an example of what the cleaned up code looks like.
I don't want this merged into the master branch, as I'm not a student.

I'll be making a robot code project in my personal GitHub using some best practices later on.

Congratulations on another season!